refactor: 設定コンポーネントを共通化
