+++ packages/shared/src/components/index.ts
@@ -1,4 +1,5 @@
 export * from "./CharacterCounter";
 export * from "./enhancers";
 export * from "./schedule";
+export * from "./settings";
 export * from "./ui";
