+++ packages/shared/src/components/settings/index.ts
@@ -1,0 +1,3 @@
+export { ThemeSettings } from "./ThemeSettings";
+export { ShortcutSettings } from "./KeyboardShortcutSettings";
+export { SettingsPage } from "./SettingsPage";
